# docs/framework/contracts/openapi-transactional.yaml
openapi: 3.1.0
info:
  title: WSDenatran REST API (mock) — Transactional RENACH/RENAINF
  version: 1.0.0
  license:
    name: UNLICENSED
    url: https://example.invalid/senatran/LICENSE
  description: >-
    Dev-only mock. This specification is the transactional half of the same
    WSDenatran mock, ported from the canonical RENACH/RENAINF adapter API into
    the WSDenatran contract conventions (see
    `docs/framework/contracts/canonical-mapping.md`). It shares the exact same
    conventions as `openapi.yaml`: single `/v1` surface, `x-cpf-usuario` header,
    certificate simulation via `x-client-cert-cn`, raw JSON success bodies and
    the shared `ErrorResponse` (`returnCode`, `message`) error envelope.


    Unlike the read-only `openapi.yaml`, these operations are transactional:
    they create and mutate RENACH driver-licensing processes and RENAINF
    administrative-infraction cases, driving their respective state machines.
    Business-rule / state violations return 402 with the canonical error code
    carried in `message` as `CODE — descrição`. Not for production use.
servers:
  - url: /v1
security:
  - clientCertSimulation: []
...
paths:
...
  /renach/processos:
    post:
      tags: [renachProcessos]
      operationId: abrirProcessoRenach
      summary: Abrir processo RENACH
      description: >-
        Abre um novo processo RENACH para o candidato e retorna o
        `numeroRenach` atribuído. Extensão WSDenatran (sem equivalente no corpus
        canônico); ver canonical-mapping.md.
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: '#/components/schemas/AberturaProcessoRenachRequest' }
      responses:
        '201':
          description: Criado. Processo RENACH aberto em situação ABERTO.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessoRenach' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.ALREADY_OPEN` — já existe processo
            ativo para o CPF e tipo informados.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}:
    get:
      tags: [renachProcessos]
      operationId: getProcessoRenach
      summary: Consultar processo RENACH
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      responses:
        '200':
          description: OK. Tudo funcionou como esperado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessoRenach' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/BusinessError' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/elegibilidadeExame:
    post:
      tags: [renachProcessos]
      operationId: verificarElegibilidadeExame
      summary: Consultar elegibilidade para exame
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ElegibilidadeExameRequest' }
      responses:
        '200':
          description: OK. Tudo funcionou como esperado.
          content:
            application/json:
              schema:
                { $ref: '#/components/schemas/ElegibilidadeExameResponse' }
        '400':
          description: >-
            Requisição inválida. `RENACH.EXAM.INVALID_TYPE`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.INVALID_STATUS`,
            `RENACH.PROCESS.BLOCKED`, `RENACH.PROCESS.EXPIRED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/clinicasCredenciadas:
    get:
      tags: [renachClinicas]
      operationId: listarClinicasCredenciadas
      summary: Listar clínicas credenciadas
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
      responses:
        '200':
          description: OK. Tudo funcionou como esperado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicaCredenciadaList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/BusinessError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/profissionaisCredenciados:
    get:
      tags: [renachClinicas]
      operationId: listarProfissionaisCredenciados
      summary: Listar profissionais credenciados
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
      responses:
        '200':
          description: OK. Tudo funcionou como esperado.
          content:
            application/json:
              schema:
                { $ref: '#/components/schemas/ProfissionalCredenciadoList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.EXAMINER.NOT_ACCREDITED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/agendamentos:
    post:
      tags: [renachAgendamentos]
      operationId: criarAgendamentoExame
      summary: Criar agendamento de exame
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgendamentoRequest' }
      responses:
        '201':
          description: Criado. Agendamento registrado com sucesso.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agendamento' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.INVALID_STATUS`,
            `RENACH.CLINIC.NOT_ACCREDITED`, `RENACH.CLINIC.INACTIVE`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/agendamentos/{idAgendamento}:
    patch:
      tags: [renachAgendamentos]
      operationId: alterarAgendamento
      summary: Cancelar ou remarcar agendamento
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: idAgendamento
          in: path
          required: true
          description: Identificador do agendamento
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgendamentoPatchRequest' }
      responses:
        '200':
          description: OK. Agendamento atualizado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agendamento' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.INVALID_STATUS`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/agendamentos/{idAgendamento}/checkin:
    post:
      tags: [renachAgendamentos]
      operationId: iniciarAtendimento
      summary: Iniciar atendimento
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: idAgendamento
          in: path
          required: true
          description: Identificador do agendamento
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckinRequest' }
      responses:
        '200':
          description: OK. Atendimento iniciado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agendamento' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.INVALID_STATUS`,
            `RENACH.BIOMETRY.NOT_MATCHED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/examesMedicos:
    post:
      tags: [renachExames]
      operationId: registrarExameMedico
      summary: Registrar exame médico
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExameMedicoRequest' }
      responses:
        '201':
          description: Criado. Exame médico registrado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Exame' }
        '400':
          description: >-
            Requisição inválida. `RENACH.EXAM.INVALID_TYPE`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.EXAM.ALREADY_RECORDED`,
            `RENACH.EXAM.RESULT_NOT_ALLOWED`, `RENACH.EXAM.REQUIRES_BOARD`,
            `RENACH.CLINIC.NOT_ACCREDITED`, `RENACH.EXAMINER.NOT_ACCREDITED`,
            `RENACH.EXAMINER.NOT_LINKED_TO_CLINIC`, `RENACH.SIGNATURE.INVALID`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/avaliacoesPsicologicas:
    post:
      tags: [renachExames]
      operationId: registrarAvaliacaoPsicologica
      summary: Registrar avaliação psicológica
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AvaliacaoPsicologicaRequest' }
      responses:
        '201':
          description: Criado. Avaliação psicológica registrada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Exame' }
        '400':
          description: >-
            Requisição inválida. `RENACH.EXAM.INVALID_TYPE`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.EXAM.ALREADY_RECORDED`,
            `RENACH.EXAM.RESULT_NOT_ALLOWED`, `RENACH.EXAMINER.NOT_ACCREDITED`,
            `RENACH.EXAMINER.NOT_LINKED_TO_CLINIC`, `RENACH.SIGNATURE.INVALID`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/exames:
    get:
      tags: [renachExames]
      operationId: consultarResultadosExames
      summary: Consultar resultados de exames
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      responses:
        '200':
          description: OK. Tudo funcionou como esperado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExameList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/BusinessError' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/exames/{idExame}/retificacoes:
    post:
      tags: [renachExames]
      operationId: retificarExame
      summary: Retificar exame
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: idExame
          in: path
          required: true
          description: Identificador do exame
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RetificacaoExameRequest' }
      responses:
        '202':
          description: Aceito. Retificação em processamento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AceiteAssincrono' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.RECTIFICATION.NOT_ALLOWED`,
            `RENACH.EXAM.RESULT_NOT_ALLOWED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/encaminhamentosJunta:
    post:
      tags: [renachJunta]
      operationId: encaminharJuntaMedica
      summary: Encaminhar para junta médica
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EncaminhamentoJuntaRequest' }
      responses:
        '202':
          description: Aceito. Encaminhamento à junta em processamento.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AceiteAssincrono' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.EXAM.REQUIRES_BOARD`,
            `RENACH.PROCESS.INVALID_STATUS`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }
  /renach/processos/{numeroRenach}/pareceresJunta:
    post:
      tags: [renachJunta]
      operationId: registrarParecerJunta
      summary: Registrar parecer de junta
      parameters:
        - $ref: '#/components/parameters/XCpfUsuario'
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: numeroRenach
          in: path
          required: true
          description: Número do processo RENACH
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ParecerJuntaRequest' }
      responses:
        '201':
          description: Criado. Parecer de junta registrado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Exame' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402':
          description: >-
            Erro de negócio. `RENACH.PROCESS.INVALID_STATUS`,
            `RENACH.EXAM.RESULT_NOT_ALLOWED`, `RENACH.SIGNATURE.INVALID`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: >-
            Não encontrado. `RENACH.PROCESS.NOT_FOUND`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ---------------------------------------------------------------------------
  # RENAINF
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    clientCertificate:
      type: mutualTLS
      description: >-
        Real-service authentication: client digital certificate (mTLS). Using the
        service requires prior registration of the certificate. Documented for
        fidelity; the mock does not perform a TLS handshake.
    clientCertSimulation:
      type: apiKey
      in: header
      name: x-client-cert-cn
      description: >-
        Mock stand-in for the real mTLS certificate. Send the allowlisted
        certificate Common Name in the `x-client-cert-cn` header. Enforced when
        `AUTH_CERT_SIMULATION=on` (default); ignored when off. Every request must
        also send `x-cpf-usuario`. See `docs/framework/contracts/auth.md`.
  parameters:
    XCpfUsuario:
      name: x-cpf-usuario
      in: header
      required: true
      description: CPF do usuário que está fazendo a requisição
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        Chave de idempotência opcional. Se enviada, é honrada para deduplicar
        escritas; caso contrário, a deduplicação usa chaves naturais
        (`numeroAit`, `numeroRenach`).
      schema:
        type: string
  responses:
    BadRequest:
      description: Requisição inválida. A requisição não foi aceita, geralmente por falta de um parâmetro requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Não autorizado. Problemas durante a autenticação do certificado ou do CPF do usuário.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessError:
      description: Requisição falhou. Os parâmetros foram validados, mas houve algum erro de negócio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Não encontrado. O recurso solicitado não existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Erro no servidor. Ocorreu algum erro interno.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        returnCode:
          type: integer
        message:
          type: string

    # -----------------------------------------------------------------------
    # Enums e tipos compartilhados
    # -----------------------------------------------------------------------
...
    AberturaProcessoRenachRequest:
      type: object
      description: Dados para abrir um novo processo RENACH.
      required: [cpf, tipoProcesso]
      properties:
        cpf: { type: string }
        tipoProcesso: { $ref: '#/components/schemas/TipoProcesso' }
        categoriaAtual: { type: string }
        categoriaPretendida: { type: string }
      example:
        cpf: '52998224725'
        tipoProcesso: RENOVACAO
        categoriaAtual: B
        categoriaPretendida: B
    ElegibilidadeExameRequest:
      type: object
      description: Dados para verificação de elegibilidade ao exame.
      required: [tipoProcesso]
      properties:
        tipoProcesso: { $ref: '#/components/schemas/TipoProcesso' }
        categoriaAtual: { type: string }
        categoriaPretendida: { type: string }
    AgendamentoRequest:
      type: object
      required: [dataHora, codigoClinica, cpfCondutor]
      properties:
        idAgendamento: { type: string }
        dataHora: { type: string, format: date-time }
        codigoClinica: { type: string }
        cpfCondutor: { type: string }
    AgendamentoPatchRequest:
      type: object
      description: Cancelamento ou remarcação de agendamento.
      properties:
        dataHora: { type: string, format: date-time }
        codigoClinica: { type: string }
        cancelar: { type: boolean }
        motivo: { type: string }
    CheckinRequest:
      type: object
      required: [cpfCondutor]
      properties:
        cpfCondutor: { type: string }
        dataHora: { type: string, format: date-time }
        biometriaValidada: { type: boolean }
    ExameMedicoRequest:
      type: object
      description: Registro de exame médico (portado de MedicalExamRequest).
      required: [idAgendamento, clinica, examinador, condutor, processo, exame]
      properties:
        idAgendamento: { type: string }
        clinica:
          type: object
          required: [codigoClinica, cnpj]
          properties:
            codigoClinica: { type: string }
            cnpj: { type: string }
        examinador:
          type: object
          required: [cpf, conselho, numeroConselho, uf]
          properties:
            cpf: { type: string }
            conselho:
              type: string
              enum: [CRM]
            numeroConselho: { type: string }
            uf: { type: string }
        condutor:
          type: object
          required: [cpf, nome, dataNascimento]
          properties:
            cpf: { type: string }
            nome: { type: string }
            dataNascimento: { type: string, format: date }
        processo:
          type: object
          required: [numeroRenach, tipoProcesso]
          properties:
            numeroRenach: { type: string }
            tipoProcesso: { $ref: '#/components/schemas/TipoProcesso' }
            categoriaAtual: { type: string }
            categoriaPretendida: { type: string }
        exame:
          type: object
          required: [dataRealizacao, resultado]
          properties:
            dataRealizacao: { type: string, format: date-time }
            resultado: { $ref: '#/components/schemas/ResultadoExame' }
            dataValidade: { type: string, format: date }
            restricoes:
              type: array
              items: { $ref: '#/components/schemas/Restricao' }
            observacoes: { type: string }
        anexos:
          type: array
          items: { $ref: '#/components/schemas/Anexo' }
        assinaturaDigital: { $ref: '#/components/schemas/AssinaturaDigital' }
    AvaliacaoPsicologicaRequest:
      type: object
      description: Registro de avaliação psicológica (portado de PsychologicalEvaluationRequest).
      required: [idAgendamento, clinica, examinador, avaliacao]
      properties:
        idAgendamento: { type: string }
        clinica:
          type: object
          properties:
            codigoClinica: { type: string }
            cnpj: { type: string }
        examinador:
          type: object
          properties:
            cpf: { type: string }
            conselho:
              type: string
              enum: [CRP]
            numeroConselho: { type: string }
            uf: { type: string }
        avaliacao:
          type: object
          required: [dataRealizacao, resultado]
          properties:
            dataRealizacao: { type: string, format: date-time }
            resultado: { $ref: '#/components/schemas/ResultadoExame' }
            dataValidade: { type: string, format: date }
            bateriaTestes:
              type: array
              items: { type: string }
            observacoes: { type: string }
        assinaturaDigital: { $ref: '#/components/schemas/AssinaturaDigital' }
...
    EncaminhamentoJuntaRequest:
      type: object
      description: Encaminhamento do processo à junta médica.
      required: [motivo]
      properties:
        motivo: { type: string }
        idExame: { type: string }
        observacoes: { type: string }
...
    ProcessoRenach:
      type: object
      properties:
        protocolo: { type: string }
        numeroRenach: { type: string }
        situacao: { $ref: '#/components/schemas/SituacaoRenach' }
        tipoProcesso: { $ref: '#/components/schemas/TipoProcesso' }
        categoriaAtual: { type: string }
        categoriaPretendida: { type: string }
        condutor:
          type: object
          properties:
            cpf: { type: string }
            nome: { type: string }
            dataNascimento: { type: string, format: date }
      example:
        protocolo: RENACH-2026-000123
        numeroRenach: RS123456789
        situacao: AGUARDANDO_MEDICO
        tipoProcesso: RENOVACAO
        categoriaAtual: B
        categoriaPretendida: B
        condutor:
          cpf: '52998224725'
          nome: MARIA SILVA SANTOS
          dataNascimento: '1985-03-14'
    ElegibilidadeExameResponse:
      type: object
      properties:
        protocolo: { type: string }
        numeroRenach: { type: string }
        situacao: { $ref: '#/components/schemas/SituacaoRenach' }
        elegivel: { type: boolean }
        motivo: { type: string }
    Agendamento:
      type: object
      properties:
        protocolo: { type: string }
        idAgendamento: { type: string }
        numeroRenach: { type: string }
        situacao: { $ref: '#/components/schemas/SituacaoRenach' }
        dataHora: { type: string, format: date-time }
        codigoClinica: { type: string }
        cpfCondutor: { type: string }
    Exame:
      type: object
      properties:
        protocolo: { type: string }
        idExame: { type: string }
        numeroRenach: { type: string }
        situacao: { $ref: '#/components/schemas/SituacaoRenach' }
        resultado: { $ref: '#/components/schemas/ResultadoExame' }
        dataRealizacao: { type: string, format: date-time }
        dataValidade: { type: string, format: date }
        restricoes:
          type: array
          items: { $ref: '#/components/schemas/Restricao' }
      example:
        protocolo: RENACH-EXAME-2026-000045
        idExame: 6f1c2a80-0000-4000-8000-000000000045
        numeroRenach: RS123456789
        situacao: EXAME_REGISTRADO
        resultado: APTO
        dataRealizacao: '2026-07-10T09:30:00.000Z'
        dataValidade: '2031-07-10'
        restricoes: []
    ExameList:
      type: object
      properties:
        exames:
          type: array
          items: { $ref: '#/components/schemas/Exame' }
...
    ClinicaCredenciadaList:
      type: object
      properties:
        clinicas:
          type: array
          items: { $ref: '#/components/schemas/ClinicaCredenciada' }
...
    ProfissionalCredenciadoList:
      type: object
      properties:
        profissionais:
          type: array
          items: { $ref: '#/components/schemas/ProfissionalCredenciado' }

    # -----------------------------------------------------------------------
    # RENAINF — request schemas
    # -----------------------------------------------------------------------
...